pyproject.toml
[project]
name = "timer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.1.1",
    "pyobjc-framework-cocoa>=12.1",
]
